Added new screens to view topic and instructor history. bugid: 5821 * Manage Topic screen:   - linked to from the Manage Topics screen.   - displays list of sessions, past and present for that topic, along with site, instructors, and reservation count   - instructors link to the new users/show page * Users/show:   - Linked to from Manage Topic screen and Reservations screen   - Displays list of sessions, past and present, that the user is an instructor for   - Sessions are grouped by topic. Topics are listed in alphabetical order. * Reservations (modified existing screen):   - List of next 5 upcoming sessions that the user is an instructor for.   - Link to Users/show screen in order to view all sessions, past and present.
In working on this, some CSS tweaks and DB performance tweaks snuck in. More will come in later commits.

Of note: Added in memory caching of Reservation counts. This may be modified when proper caching is implemented.
Also: Changed "Certificate of Completion" link to "Download Certificate" and added an icon. Also added icon to Survey Link.